+++ src/api/getDepartmentOrCategory.ts
@@ -4,6 +4,13 @@ import type { Filters } from 'src/components/search/PLPFilters/Filters'
 
 interface RequestType {
   slug: string
+  sort:
+    | 'Price/BasePrice asc'
+    | 'Price/BasePrice desc'
+    | 'Rank/Score desc'
+    | 'Name asc'
+    | 'Name desc'
+    | 'ReleaseDate desc'
   filteredFacets?: Filters[]
 }
 
@@ -18,10 +25,11 @@ export default async function getDepartmentOrCategory(
     return
   }
 
-  const { slug, filteredFacets } = req.body
+  const { slug, filteredFacets, sort } = req.body
 
   const categorySlug = slug.includes('/') && slug.split('/')[1]
   const departmentSlug = !slug.includes('/') && slug
+  const orderSort = sort ? `&$orderby=${sort}` : ''
 
   const filterParam = !categorySlug
     ? `Department/Slug eq '${departmentSlug}'`
@@ -86,7 +94,7 @@ export default async function getDepartmentOrCategory(
   }
   `
 
-  const URL = `${process.env.GATSBY_CATALOG_BASE_URL}/odata/Catalog/v1/Products?$expand=SKU, Department, Category, Price, Checkout, Especificacao, CommercialCondition, TradePolicy, Stock, Rank, Brand, Especificacao/CargaHoraria&$orderby=Rank/Score desc&$skip=0&$filter=${allFilters} &$top=20&$count=true&$select=ID, Name, ProductImageURL, Price/BasePrice, Price/ListPrice, Price/CommisionedPrice, Price/isSale, Category/Name, Category/Slug, Especificacao/CargaHoraria/Text, LinkId`
+  const URL = `${process.env.GATSBY_CATALOG_BASE_URL}/odata/Catalog/v1/Products?$expand=SKU, Department, Category, Price, Checkout, Especificacao, CommercialCondition, TradePolicy, Stock, Rank, Brand, Especificacao/CargaHoraria&$skip=0${orderSort}&$filter=${allFilters} &$top=20&$count=true&$select=ID, Name, ProductImageURL, Price/BasePrice, Price/ListPrice, Price/CommisionedPrice, Price/isSale, Category/Name, Category/Slug, Especificacao/CargaHoraria/Text, LinkId`
 
   try {
     const { data } = await axios.get(URL)